+++ tagcloud_fixed/changelog.md
@@ -41,3 +41,4 @@ Note: in this repository because plugin name conflict needs to be solved at firs
 
 ### Updated
 - SEO URLs. Old way is saved with new e_url addons. Fully configurable like before. No htaccess is needed. 
+- global shortcodes workaround. e_shortcode addon used.